+++ CHANGELOG.md
@@ -13,6 +13,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Show the main contact for a project in the project CSV exports
 - The all conditions met task for a transfer has updated content.
 - Add deeds of termination for MFA task to Transfer projects
+- Added the 'Deed of termination for the church supplemental agreement' task to
+  transfers.
 
 ### Changed